Getting window at index zero, to avoid cases when the keywindow is a UIAlertview's superview.
To replicate the issue, call the authorize method from "- (void)alertView:(UIAlertView *)alertView clickedButtonAtIndex:(NSInteger)buttonIndex" the FB dialog will sometimes appear for few moments and then it will disappear. When debugging I could see that the FBDialog was getting added to the window where the alert view is presented (_UIAlertOverlayWindow) that is the keyWindow.